add shared File user info

+++ core/server/master/src/main/java/alluxio/master/file/DefaultFileSystemMaster.java
@@ -1292,7 +1292,8 @@ public final class DefaultFileSystemMaster extends AbstractMaster implements Fil
       deletedInodes = deleteAndJournal(inodePath, options, journalContext);
       auditContext.setSucceeded(true);
       //add by li;
-      mInodeTree.deleteFromUser(inodePath);
+      if(inodePath.getInode().isFile())
+        mInodeTree.deleteFromUser(inodePath);
     }
     deleteInodeBlocks(deletedInodes);
   }
@@ -3096,16 +3097,36 @@ public final class DefaultFileSystemMaster extends AbstractMaster implements Fil
     }
   }
 
-  //add by li
-  public Set<Inode> getInfoByUser(String owner) {
-    return mInodeTree.getInfoByUser(owner);
-  }
-
   @Override
   public List<WorkerInfo> getWorkerInfoList() {
     return mBlockMaster.getWorkerInfoList();
   }
 
+  //=======================================add by li==========================================
+  //TODO(li) make it safe
+  @Override
+  public Set<Inode> getInfoByUser(String owner) {
+    return mInodeTree.getInfoByUser(owner);
+  }
+
+  @Override
+  public List<String> getUsersByUrl(AlluxioURI uri) throws InvalidPathException,
+      FileDoesNotExistException{
+    //only handle inodeFile
+    List<String> res = new ArrayList<>();
+    try (LockedInodePath inodePath = mInodeTree.lockInodePath(uri, InodeTree.LockMode.WRITE)){
+      res.addAll(inodePath.getInode().getOwners());
+      res.add(inodePath.getInode().getOwner());
+      return res;
+    }
+  }
+
+  @Override
+  public void addUser(String user, long fileId)  throws FileDoesNotExistException{
+    mInodeTree.addUser(user, fileId);
+
+  }
+  //=======================================add by li==========================================
   /**
    * Class that contains metrics for FileSystemMaster.
    * This class is public because the counter names are referenced in
@@ -3142,6 +3163,8 @@ public final class DefaultFileSystemMaster extends AbstractMaster implements Fil
     private static final Counter RENAME_PATH_OPS = MetricsSystem.masterCounter("RenamePathOps");
     private static final Counter SET_ATTRIBUTE_OPS = MetricsSystem.masterCounter("SetAttributeOps");
     private static final Counter UNMOUNT_OPS = MetricsSystem.masterCounter("UnmountOps");
+    //add by li
+    //private static final Counter ADD_USER_OPTS = MetricsSystem.masterCounter("addUserOpts");
 
     public static final String FILES_PINNED = "FilesPinned";
     public static final String PATHS_TOTAL = "PathsTotal";
